refactor: move SpinalSimStreamFifoCCTester
